Clean class location
