+++ Sources/vbSparkle.Console/Program.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using vbSparkle.Options;
 
 namespace vbSparkle.CLI
 {
