+++ Sources/vbSparkle/LanguageStatements/Blocks/VbIfThenElseStatement.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Text;
+using vbSparkle.Options;
 using static vbSparkle.VBScriptParser;
 
 namespace vbSparkle
