+++ Sources/vbSparkle/LanguageStatements/UserScopeObjects/VbUserScopeObject.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Antlr4.Runtime.Tree;
 using System.Collections.Generic;
+using vbSparkle.Options;
 
 namespace vbSparkle
 {
