+++ Sources/vbSparkle/LanguageStatements/VbModule.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using vbSparkle.Options;
 
 namespace vbSparkle
 {
