+++ Sources/vbSparkle/Options/JunkCodeProcessingMode.cs
@@ -1,0 +1,9 @@
+﻿namespace vbSparkle.Options
+{
+    public enum JunkCodeProcessingMode
+    {
+        Nothing = 0,
+        Remove = 1,
+        Comment = 2
+    }
+}
