+++ Sources/vbSparkle/Options/SymbolRenamingMode.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace vbSparkle.Options
+{
+    [Flags]
+    public enum SymbolRenamingMode
+    {
+        None,
+        Variables,
+        Constants,
+        PublicMembers,
+        PrivateMembers,
+        Members = PublicMembers | PrivateMembers,
+        All = Variables | Constants | Members,
+        AutoDetectObfuscatedSymbols,
+
+    }
+}
