+++ Sources/vbSparkle/VbPreProcessorsAnalyser.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using vbSparkle.Options;
 using static vbSparkle.VBPreprocessorsParser;
 
 namespace vbSparkle
